+++ HISTORY.md
@@ -9,6 +9,7 @@
     anymore because of better override
 * extensions: added html extensions to add elements to another and
     get the body or head element of the HTML frame
+* extensions: htm_text to get/set html element's text
 
 ### 2025-11-23
 * basic HTML support!